move form handling to ContactForm.js handleSubmitForm function
